Convert server tests to TSX
